requirements.txt
aiocache==0.12.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncio-throttle==1.0.2
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==4.2.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cryptography==43.0.0
Deprecated==1.2.14
fastapi==0.112.1
frozenlist==1.4.1
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
limits==3.13.0
locust==2.31.5
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.0.8
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pylogbeat==2.0.1
PyMySQL==1.1.1
pytest==8.3.2
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-logstash==0.4.8
python-logstash-async==3.0.0
python-telegram-bot==21.5
pytz==2024.1
pyzmq==26.2.0
ratelimit==2.2.1
redis==5.0.8
requests==2.32.3
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
spotipy==2.19.0
SQLAlchemy==1.4.31
starlette==0.38.2
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.15
uvicorn==0.30.6
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.0
zope.event==5.0
zope.interface==7.0.3